move method orientaitons and minimum_outdegree_orientation
