+++ src/sage/graphs/orientations.py
@@ -13,12 +13,13 @@ etc.). It also implements some iterators over all these orientations.
     :delim: |
 
     :meth:`orient` | Return an oriented version of `G` according the input function `f`.
+    :meth:`orientations` | Return an iterator over orientations of `G`.
     :meth:`acyclic_orientations` | Return an iterator over all acyclic orientations of an undirected graph `G`.
     :meth:`strong_orientation` | Return a strongly connected orientation of the graph `G`.
     :meth:`strong_orientations_iterator` | Return an iterator over all strong orientations of a graph `G`
     :meth:`random_orientation` | Return a random orientation of a graph `G`
     :meth:`minimum_outdegree_orientation` | Return an orientation of `G` with the smallest possible maximum outdegree.
-
+    :meth:`bounded_outdegree_orientation` | Return an orientation of `G` such that every vertex `v` has out-degree less than `b(v)`.
 
 Authors
 -------
@@ -220,6 +221,132 @@ def orient(G, f, weighted=None, data_structure=None, sparse=None,
                 setattr(D, attr, copy(old_attr))
 
     return D
+
+
+def orientations(G, data_structure=None, sparse=None):
+    r"""
+    Return an iterator over orientations of `G`.
+
+    An *orientation* of an undirected graph is a directed graph such that
+    every edge is assigned a direction.  Hence there are `2^s` oriented
+    digraphs for a simple graph with `s` edges.
+
+    INPUT:
+
+    - ``data_structure`` -- one of ``'sparse'``, ``'static_sparse'``, or
+      ``'dense'``; see the documentation of :class:`Graph` or :class:`DiGraph`;
+      default is the data structure of `G`
+
+    - ``sparse`` -- boolean (default: ``None``); ``sparse=True`` is an alias for
+      ``data_structure="sparse"``, and ``sparse=False`` is an alias for
+      ``data_structure="dense"``. By default (``None``), guess the most suitable
+      data structure.
+
+    .. WARNING::
+
+        This always considers multiple edges of graphs as distinguishable, and
+        hence, may have repeated digraphs.
+
+    .. SEEALSO::
+
+        - :meth:`~sage.graphs.graph.Graph.strong_orientation`
+        - :meth:`~sage.graphs.orientations.strong_orientations_iterator`
+        - :meth:`~sage.graphs.digraph_generators.DiGraphGenerators.nauty_directg`
+        - :meth:`~sage.graphs.orientations.random_orientation`
+
+    EXAMPLES::
+
+        sage: G = Graph([[1,2,3], [(1, 2, 'a'), (1, 3, 'b')]], format='vertices_and_edges')
+        sage: it = G.orientations()
+        sage: D = next(it)
+        sage: D.edges(sort=True)
+        [(1, 2, 'a'), (1, 3, 'b')]
+        sage: D = next(it)
+        sage: D.edges(sort=True)
+        [(1, 2, 'a'), (3, 1, 'b')]
+
+    TESTS::
+
+        sage: G = Graph()
+        sage: D = [g for g in G.orientations()]
+        sage: len(D)
+        1
+        sage: D[0]
+        Digraph on 0 vertices
+
+        sage: G = Graph(5)
+        sage: it = G.orientations()
+        sage: D = next(it)
+        sage: D.size()
+        0
+
+        sage: G = Graph([[1,2,'a'], [1,2,'b']], multiedges=True)
+        sage: len(list(G.orientations()))
+        4
+
+        sage: G = Graph([[1,2], [1,1]], loops=True)
+        sage: len(list(G.orientations()))
+        2
+
+        sage: G = Graph([[1,2],[2,3]])
+        sage: next(G.orientations())
+        Digraph on 3 vertices
+        sage: G = graphs.PetersenGraph()
+        sage: next(G.orientations())
+        An orientation of Petersen graph: Digraph on 10 vertices
+
+    An orientation must have the same ground set of vertices as the original
+    graph (:issue:`24366`)::
+
+        sage: G = Graph(1)
+        sage: next(G.orientations())
+        Digraph on 1 vertex
+    """
+    if sparse is not None:
+        if data_structure is not None:
+            raise ValueError("cannot specify both 'sparse' and 'data_structure'")
+        data_structure = "sparse" if sparse else "dense"
+    if data_structure is None:
+        from sage.graphs.base.dense_graph import DenseGraphBackend
+        from sage.graphs.base.sparse_graph import SparseGraphBackend
+        if isinstance(G._backend, DenseGraphBackend):
+            data_structure = "dense"
+        elif isinstance(G._backend, SparseGraphBackend):
+            data_structure = "sparse"
+        else:
+            data_structure = "static_sparse"
+
+    name = G.name()
+    if name:
+        name = 'An orientation of ' + name
+
+    if not G.size():
+        D = DiGraph(data=[G, []],
+                    format='vertices_and_edges',
+                    name=name,
+                    pos=G._pos,
+                    multiedges=G.allows_multiple_edges(),
+                    loops=G.allows_loops(),
+                    data_structure=data_structure)
+        if hasattr(G, '_embedding'):
+            D._embedding = copy(G._embedding)
+        yield D
+        return
+
+    E = [[(u, v, label), (v, u, label)] if u != v else [(u, v, label)]
+         for u, v, label in G.edge_iterator()]
+    from itertools import product
+    for edges in product(*E):
+        D = DiGraph(data=[G, edges],
+                    format='vertices_and_edges',
+                    name=name,
+                    pos=G._pos,
+                    multiedges=G.allows_multiple_edges(),
+                    loops=G.allows_loops(),
+                    data_structure=data_structure)
+        if hasattr(G, '_embedding'):
+            D._embedding = copy(G._embedding)
+        yield D
 
 
 def acyclic_orientations(G):
@@ -972,3 +1099,166 @@ def minimum_outdegree_orientation(G, use_edge_labels=False, solver=None, verbose
     O.delete_edges(edges)
 
     return O
+
+
+def bounded_outdegree_orientation(G, bound, solver=None, verbose=False,
+                                  *, integrality_tolerance=1e-3):
+    r"""
+    Return an orientation of `G` such that every vertex `v` has out-degree less
+    than `b(v)`.
+
+    INPUT:
+
+    - ``bound`` -- maximum bound on the out-degree. Can be of three
+      different types :
+
+      * An integer `k`. In this case, computes an orientation whose maximum
+        out-degree is less than `k`.
+
+      * A dictionary associating to each vertex its associated maximum
+        out-degree.
+
+      * A function associating to each vertex its associated maximum
+        out-degree.
+
+    - ``solver`` -- string (default: ``None``); specifies a Mixed Integer Linear
+      Programming (MILP) solver to be used. If set to ``None``, the default one
+      is used. For more information on MILP solvers and which default solver is
+      used, see the method :meth:`solve
+      <sage.numerical.mip.MixedIntegerLinearProgram.solve>` of the class
+      :class:`MixedIntegerLinearProgram
+      <sage.numerical.mip.MixedIntegerLinearProgram>`.
+
+    - ``verbose`` -- integer (default: 0); sets the level of verbosity. Set to 0
+      by default, which means quiet.
+
+    - ``integrality_tolerance`` -- float; parameter for use with MILP solvers
+      over an inexact base ring;
+      see :meth:`MixedIntegerLinearProgram.get_values`.
+
+    OUTPUT:
+
+    A DiGraph representing the orientation if it exists.
+    A :exc:`ValueError` exception is raised otherwise.
+
+    ALGORITHM:
+
+    The problem is solved through a maximum flow :
+
+    Given a graph `G`, we create a ``DiGraph`` `D` defined on `E(G)\cup V(G)\cup
+    \{s,t\}`. We then link `s` to all of `V(G)` (these edges having a capacity
+    equal to the bound associated to each element of `V(G)`), and all the
+    elements of `E(G)` to `t` . We then link each `v \in V(G)` to each of its
+    incident edges in `G`. A maximum integer flow of value `|E(G)|` corresponds
+    to an admissible orientation of `G`. Otherwise, none exists.
+
+    EXAMPLES:
+
+    There is always an orientation of a graph `G` such that a vertex `v` has
+    out-degree at most `\lceil \frac {d(v)} 2 \rceil`::
+
+        sage: g = graphs.RandomGNP(40, .4)
+        sage: b = lambda v: integer_ceil(g.degree(v)/2)
+        sage: D = g.bounded_outdegree_orientation(b)
+        sage: all( D.out_degree(v) <= b(v) for v in g )
+        True
+
+    Chvatal's graph, being 4-regular, can be oriented in such a way that its
+    maximum out-degree is 2::
+
+        sage: g = graphs.ChvatalGraph()
+        sage: D = g.bounded_outdegree_orientation(2)
+        sage: max(D.out_degree())
+        2
+
+    For any graph `G`, it is possible to compute an orientation such that
+    the maximum out-degree is at most the maximum average degree of `G`
+    divided by 2. Anything less, though, is impossible.
+
+        sage: g = graphs.RandomGNP(40, .4)
+        sage: mad = g.maximum_average_degree()                                      # needs sage.numerical.mip
+
+    Hence this is possible ::
+
+        sage: d = g.bounded_outdegree_orientation(integer_ceil(mad/2))              # needs sage.numerical.mip
+
+    While this is not::
+
+        sage: try:                                                                  # needs sage.numerical.mip
+        ....:     g.bounded_outdegree_orientation(integer_ceil(mad/2-1))
+        ....:     print("Error")
+        ....: except ValueError:
+        ....:     pass
+
+    TESTS:
+
+    As previously for random graphs, but more intensively::
+
+        sage: for i in range(30):      # long time (up to 6s on sage.math, 2012)
+        ....:     g = graphs.RandomGNP(40, .4)
+        ....:     b = lambda v: integer_ceil(g.degree(v)/2)
+        ....:     D = g.bounded_outdegree_orientation(b)
+        ....:     if not (
+        ....:          all( D.out_degree(v) <= b(v) for v in g ) or
+        ....:          D.size() != g.size()):
+        ....:         print("Something wrong happened")
+    """
+    G._scream_if_not_simple()
+    n = G.order()
+
+    if not n:
+        return DiGraph()
+
+    vertices = list(G)
+    vertices_id = {y: x for x, y in enumerate(vertices)}
+
+    b = {}
+
+    # Checking the input type. We make a dictionary out of it
+    if isinstance(bound, dict):
+        b = bound
+    else:
+        try:
+            b = dict(zip(vertices, map(bound, vertices)))
+        except TypeError:
+            b = dict(zip(vertices, [bound]*n))
+
+    d = DiGraph()
+
+    # Adding the edges (s,v) and ((u,v),t)
+    d.add_edges(('s', vertices_id[v], b[v]) for v in vertices)
+
+    d.add_edges(((vertices_id[u], vertices_id[v]), 't', 1)
+                 for u, v in G.edges(sort=False, labels=None))
+
+    # each v is linked to its incident edges
+
+    for u, v in G.edge_iterator(labels=None):
+        u, v = vertices_id[u], vertices_id[v]
+        d.add_edge(u, (u, v), 1)
+        d.add_edge(v, (u, v), 1)
+
+    # Solving the maximum flow
+    value, flow = d.flow('s', 't', value_only=False, integer=True,
+                         use_edge_labels=True, solver=solver, verbose=verbose,
+                         integrality_tolerance=integrality_tolerance)
+
+    if value != G.size():
+        raise ValueError("No orientation exists for the given bound")
+
+    D = DiGraph()
+    D.add_vertices(vertices)
+
+    # The flow graph may not contain all the vertices, if they are
+    # not part of the flow...
+
+    for u in [x for x in range(n) if x in flow]:
+
+        for uu, vv in flow.neighbors_out(u):
+            v = vv if vv != u else uu
+            D.add_edge(vertices[u], vertices[v])
+
+    # I do not like when a method destroys the embedding ;-)
+    D.set_pos(G.get_pos())
+
+    return D
